# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Paper trading system API with multi-agent AI
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List all orders
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, filled, cancelled, all]
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: placeOrder
      tags: [orders]
      summary: Place a manual trade order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/cancel:
    post:
      operationId: cancelOrder
      tags: [orders]
      summary: Cancel a pending order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
components:
  schemas:
...
    Order:
      type: object
      required: [id, symbol, assetType, side, orderType, quantity, status, createdAt]
      properties:
        id:
          type: integer
        symbol:
          type: string
        assetType:
          type: string
          enum: [stock, option]
        side:
          type: string
          enum: [buy, sell]
        orderType:
          type: string
          enum: [market, limit, stop]
        quantity:
          type: number
        limitPrice:
          type: ["number", "null"]
        stopPrice:
          type: ["number", "null"]
        filledPrice:
          type: ["number", "null"]
        status:
          type: string
          enum: [pending, filled, cancelled, rejected]
        agentId:
          type: ["integer", "null"]
        agentName:
          type: ["string", "null"]
        reason:
          type: ["string", "null"]
        optionType:
          type: ["string", "null"]
        strikePrice:
          type: ["number", "null"]
        expirationDate:
          type: ["string", "null"]
        createdAt:
          type: string
        filledAt:
          type: ["string", "null"]
    OrderInput:
      type: object
      required: [symbol, assetType, side, orderType, quantity]
      properties:
        symbol:
          type: string
        assetType:
          type: string
          enum: [stock, option]
        side:
          type: string
          enum: [buy, sell]
        orderType:
          type: string
          enum: [market, limit, stop]
        quantity:
          type: number
        limitPrice:
          type: number
        stopPrice:
          type: number
        optionType:
          type: string
          enum: [call, put]
        strikePrice:
          type: number
        expirationDate:
          type: string
...