# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Paper trading system API with multi-agent AI
 * OpenAPI spec version: 0.1.0
 */
import type { OrderAssetType } from './orderAssetType';
import type { OrderOrderType } from './orderOrderType';
import type { OrderSide } from './orderSide';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  symbol: string;
  assetType: OrderAssetType;
  side: OrderSide;
  orderType: OrderOrderType;
  quantity: number;
  /** @nullable */
  limitPrice?: number | null;
  /** @nullable */
  stopPrice?: number | null;
  /** @nullable */
  filledPrice?: number | null;
  status: OrderStatus;
  /** @nullable */
  agentId?: number | null;
  /** @nullable */
  agentName?: string | null;
  /** @nullable */
  reason?: string | null;
  /** @nullable */
  optionType?: string | null;
  /** @nullable */
  strikePrice?: number | null;
  /** @nullable */
  expirationDate?: string | null;
  createdAt: string;
  /** @nullable */
  filledAt?: string | null;
}